Move blocks/Filters to components
