+++ src/pages/index.js
@@ -8,7 +8,7 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { ClientProvider } from 'context/ClientContext'
 import get from 'utils/get'
 import Header from 'components/Header'
-import Filters from 'blocks/Filters'
+import Filters from 'components/Filters'
 import MoviesList from 'components/MoviesList/index'
 import { LANGUAGES } from 'data/constants'
 import { LanguageProvider } from 'context/LanguageContext'
